fix missing details from s3
lint
